refactor: 체스판의 초기화 시 전략패턴 사용
