+++ src/main/java/chess/domain/board/Board.java
@@ -1,17 +1,9 @@
 package chess.domain.board;
 
-import chess.domain.piece.Bishop;
 import chess.domain.piece.Direction;
-import chess.domain.piece.King;
-import chess.domain.piece.Knight;
-import chess.domain.piece.Pawn;
 import chess.domain.piece.Piece;
-import chess.domain.piece.Queen;
-import chess.domain.piece.Rook;
 import chess.domain.piece.Team;
-import chess.domain.position.Column;
 import chess.domain.position.Position;
-import chess.domain.position.Row;
 import java.util.HashMap;
 import java.util.Map;
 
@@ -20,9 +12,8 @@ public class Board {
     private final Map<Position, Piece> board = new HashMap<>();
     private Team turn = Team.WHITE;
 
-    public void initBoard() {
-        initWhitePieces();
-        initBlackPieces();
+    public void initBoard(BoardGenerator boardGenerator) {
+        board.putAll(boardGenerator.create());
     }
 
     public Map<Position, Piece> getBoard() {
@@ -56,38 +47,6 @@ public class Board {
                 throw new IllegalArgumentException("이동 경로에 말이 있습니다.");
             }
             current = current.move(direction);
-        }
-    }
-
-    private void initBlackPieces() {
-        board.put(new Position(Column.A, Row.EIGHT), new Rook(Team.BLACK));
-        board.put(new Position(Column.B, Row.EIGHT), new Knight(Team.BLACK));
-        board.put(new Position(Column.C, Row.EIGHT), new Bishop(Team.BLACK));
-        board.put(new Position(Column.D, Row.EIGHT), new Queen(Team.BLACK));
-        board.put(new Position(Column.E, Row.EIGHT), new King(Team.BLACK));
-        board.put(new Position(Column.F, Row.EIGHT), new Bishop(Team.BLACK));
-        board.put(new Position(Column.G, Row.EIGHT), new Knight(Team.BLACK));
-        board.put(new Position(Column.H, Row.EIGHT), new Rook(Team.BLACK));
-
-        initOneLine(Row.SEVEN, new Pawn(Team.BLACK));
-    }
-
-    private void initWhitePieces() {
-        board.put(new Position(Column.A, Row.ONE), new Rook(Team.WHITE));
-        board.put(new Position(Column.B, Row.ONE), new Knight(Team.WHITE));
-        board.put(new Position(Column.C, Row.ONE), new Bishop(Team.WHITE));
-        board.put(new Position(Column.D, Row.ONE), new Queen(Team.WHITE));
-        board.put(new Position(Column.E, Row.ONE), new King(Team.WHITE));
-        board.put(new Position(Column.F, Row.ONE), new Bishop(Team.WHITE));
-        board.put(new Position(Column.G, Row.ONE), new Knight(Team.WHITE));
-        board.put(new Position(Column.H, Row.ONE), new Rook(Team.WHITE));
-
-        initOneLine(Row.TWO, new Pawn(Team.WHITE));
-    }
-
-    private void initOneLine(Row row, Piece piece) {
-        for (Column column : Column.values()) {
-            board.put(new Position(column, row), piece);
         }
     }
