+++ src/main/java/chess/domain/board/BoardGenerator.java
@@ -1,0 +1,9 @@
+package chess.domain.board;
+
+import chess.domain.piece.Piece;
+import chess.domain.position.Position;
+import java.util.Map;
+
+public interface BoardGenerator {
+    Map<Position, Piece> create();
+}
